+++ src/index.js
@@ -1,11 +1,12 @@
 // Authorisation
 import './js/firebase';
-// 
+//
 
 import './js/header';
 import './js/launch';
 import './js/bestsellers';
 import './js/filter';
+import './js/see-more-btn';
 import './js/categories';
 import './js/theme-dark-light';
 import './js/support/support';
@@ -18,5 +19,3 @@ import './js/modal-auth-out';
 import './js/modalcardbook';
 import './js/modal-auth';
 import { showLoader, hideLoader } from './js/loader';
-
-
