+++ src/js/filter.js
@@ -1,23 +1,64 @@
+import Notiflix from 'notiflix';
+
 import BookAPI from './book-api';
 import markupBook from './render-book-card';
-import Notiflix from 'notiflix';
 import renderWrapCategories from './bestsellers';
 import { showLoader, hideLoader } from './loader';
+import scrollToMainTitle from './see-more-btn';
 
 const bookApi = new BookAPI();
 const containerContent = document.querySelector('.books-render-js');
 const categorieEl = document.querySelector('.categorie-js');
 const homeContainerEl = document.querySelector('.home-container');
-const mainTitleEl = document.querySelector('.main-title');
 const categoriesListEl = document.querySelector('.categories_list');
-let filterItemArray = [];
+
 let activeCategoty = null;
+export let filterItemArray = [];
 
 export default function renderCategories(categories) {
   markupCategorie({ list_name: 'All categories' });
   categories
     .sort((a, b) => a.list_name.localeCompare(b.list_name))
     .map(categorie => markupCategorie(categorie));
+}
+
+export function generateCategory(name) {
+  homeContainerEl.classList.add('hidden');
+  showLoader();
+  bookApi
+    .getSelectedCategoryBooks(name)
+    .then(data => {
+      renderBooks(data);
+      renderMainTitle(name);
+      hideLoader();
+      homeContainerEl.classList.remove('hidden');
+    })
+    .catch(error => {
+      console.error('Error found category:', error);
+      Notiflix.Notify.failure(
+        'Oops, there is no category with that name. Please try again later.'
+      );
+    });
+}
+
+function generateBestSellersCategories() {
+  homeContainerEl.classList.add('hidden');
+  showLoader();
+  bookApi
+    .getTopBooks()
+    .then(data => {
+      cleaningBooks();
+      renderWrapCategories(data);
+      renderMainTitle('Best Seller Books');
+      hideLoader();
+      homeContainerEl.classList.remove('hidden');
+    })
+    .catch(error => {
+      console.error('Error retrieving top books:', error);
+      Notiflix.Notify.failure(
+        'Oops! Error retrieving top books. Please try again later.'
+      );
+    });
 }
 
 function renderBooks(books) {
@@ -50,6 +91,7 @@ function markupCategorie({ list_name }) {
 
   filterItem.addEventListener('click', () => {
     isActiveCategoryBtn(filterItem);
+    scrollToMainTitle();
     if (name === 'All categories') {
       generateBestSellersCategories();
     } else {
@@ -59,45 +101,6 @@ function markupCategorie({ list_name }) {
 
   categorieEl.appendChild(element.firstChild);
   activeCategoty = categoriesListEl.firstElementChild;
-}
-
-function generateBestSellersCategories() {
-  homeContainerEl.classList.add('hidden');
-  showLoader();
-  bookApi
-    .getTopBooks()
-    .then(data => {
-      cleaningBooks();
-      renderWrapCategories(data);
-      renderMainTitle('Best Seller Books');
-      hideLoader();
-      homeContainerEl.classList.remove('hidden');
-    })
-    .catch(error => {
-      console.error('Error retrieving top books:', error);
-      Notiflix.Notify.failure(
-        'Oops! Error retrieving top books. Please try again later.'
-      );
-    });
-}
-
-function generateCategory(name) {
-  homeContainerEl.classList.add('hidden');
-  showLoader();
-  bookApi
-    .getSelectedCategoryBooks(name)
-    .then(data => {
-      renderBooks(data);
-      renderMainTitle(name);
-      hideLoader();
-      homeContainerEl.classList.remove('hidden');
-    })
-    .catch(error => {
-      console.error('Error found category:', error);
-      Notiflix.Notify.failure(
-        'Oops, there is no category with that name. Please try again later.'
-      );
-    });
 }
 
 function cleaningBooks() {
@@ -112,7 +115,7 @@ function cleaningTitle() {
   }
 }
 
-function isActiveCategoryBtn(filterItem) {
+export function isActiveCategoryBtn(filterItem) {
   if (activeCategoty) {
     activeCategoty.classList.remove('filter-item-is-Active');
   }
@@ -120,23 +123,35 @@ function isActiveCategoryBtn(filterItem) {
   activeCategoty = filterItem;
 }
 
-containerContent.addEventListener('click', function (event) {
-  if (event.target.classList.contains('see-more-btn')) {
-    const listName = event.target.dataset.active;
-    window.scrollTo({ top: mainTitleEl.offsetTop, behavior: 'smooth' });
-    generateCategory(listName);
-    searchCategoryName(listName);
-  }
-});
+// function onClickSeeMoreBtn(event) {
+//   if (event.target.classList.contains('see-more-btn')) {
+//     const listName = event.target.dataset.active;
+//     generateCategory(listName);
+//     searchCategoryName(listName);
+//     let elementFound = searchCategoryName(listName);
+//     // scrollToElement(elementFound);
+//     scrollToMainTitle();
+//   }
+// }
 
-function searchCategoryName(list_name) {
-  filterItemArray.forEach(element => {
-    if (list_name === element.dataset.active) {
-      isActiveCategoryBtn(element);
-    }
-  });
-}
+// containerContent.addEventListener('click', onClickSeeMoreBtn);
 
+// function searchCategoryName(list_name) {
+//   let foundElement = null;
+//   filterItemArray.forEach(element => {
+//     if (list_name === element.dataset.active) {
+//       isActiveCategoryBtn(element);
+//       foundElement = element;
+//       return;
+//     }
+//   });
+//   return foundElement;
+// }
 
+// function scrollToElement(elementToScroll) {
+//   elementToScroll.scrollIntoView({ block: 'center', behavior: 'smooth' });
+// }
 
-
+//  function scrollToMainTitle() {
+//   window.scrollTo({ top: homeContainerEl.offsetTop, behavior: 'smooth' });
+// }
